Talk_to_your_data/tools/requirements.txt
langchain==0.1.14
langchain_community==0.0.31
langchain_core==0.1.40
langchain_openai==0.1.1
python-dotenv==1.0.1
faiss-cpu
